LayoutTests/fast/encoding/stream-visibility.html

<!doctype html>
<html>
    <head>
        <meta charset="utf-8">
        <script src="../../resources/testharness.js"></script>
        <script src="../../resources/testharnessreport.js"></script>
    </head>
    <body>
        <script>
test(() => {
    assert_equals(window.TextEncoderStreamEncoder, undefined, "encoder stream");
    assert_equals(window.TextDecoderStreamDecoder, undefined, "decoder stream");
}, "Encoder/decoder exposure");
        </script>
</body>
</html>
